+++ CHANGELOG.md
@@ -4,6 +4,16 @@ earthdata-varinfo follows semantic versioning. All notable changes to this
 project will be documented in this file. The format is based on
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [v5.1.0] - 2026-05-15
+
+### Added
+
+* This version of `earthdata-varinfo` adds support for identifying explicitly
+  excluded science variables. This includes `get_excluded_science_variables()`
+  to retrieve a complete set of excluded variables defined in the JSON
+  configuration, and `is_excluded_science_variable()` to evaluate the exclusion
+  status of a specific variable.
+
 ## [v5.0.0] - 2026-05-14
 
 ### Removed
@@ -266,6 +276,8 @@ to the repository include updated documentation and files outlined by the
 For more information on internal releases prior to NASA open-source approval,
 see `legacy-CHANGELOG.md`.
 
+[v5.1.0]: https://github.com/nasa/earthdata-varinfo/releases/tag/5.1.0
+[v5.0.0]: https://github.com/nasa/earthdata-varinfo/releases/tag/5.0.0
 [v4.1.1]: https://github.com/nasa/earthdata-varinfo/releases/tag/4.1.1
 [v4.1.0]: https://github.com/nasa/earthdata-varinfo/releases/tag/4.1.0
 [v4.0.1]: https://github.com/nasa/earthdata-varinfo/releases/tag/4.0.1